[Progress: #4904862] Fellow can reorder stories + minor fixes

+++ lib/synchronizable.rb
@@ -10,7 +10,7 @@ module Synchronizable
     base.class_variable_set(:@@synchronizable_fields, [])
     base.extend ClassMethods
     base.field :_remote_id, :type => Integer
-    base.field :_collection_order, :type => Integer, :default => 9999
+    base.field :_collection_order, :type => Float, :default => 9999.0
   end
   
   module ClassMethods
